requirements.txt
fastapi==0.99
pydantic
uvicorn
mangum
python-multipart
jinja2

boto3
botocore

twilio

openai
websockets

python-dotenv
aiohttp
pydub